fix(AutoComplete): fix bugs and also fix imports in docs
