--- src/components/AutoComplete/__stand__/AutoComplete.stand.mdx
@@ -1,11 +1,0 @@
-import { AutoCompleteExampleBasic } from './examples/AutoCompleteExampleBasic/AutoCompleteExampleBasic';
-
-```tsx
-import { AutoComplete } from '@consta/uikit/AutoComplete';
-```
-
-Поле ввода с подсказками. Позволяет как написать текст вручную так и использовать
-предустановленные подсказки. Этот компонент выглядит как [TextField](##LIBS.LIB.STAND/lib:uikit/stand:components-textfield-stable),
-но возможностей больше.
-
-<AutoCompleteExampleBasic />
