--- src/components/AutoComplete/__stand__/examples/AutoCompleteExampleGroups/AutoCompleteExampleGroups.tsx
@@ -1,23 +1,0 @@
-import React, { useState } from 'react';
-
-import { groups, items } from '##/components/AutoComplete/__mocks__/data.mock';
-import { AutoComplete } from '##/components/AutoComplete/AutoComplete';
-
-import { cnDocsDecorator } from '../../../../../uiKit/components/DocsDecorator/DocsDecorator';
-import { StoryBookExample } from '../../../../../uiKit/components/StoryBookExample/StoryBookExample';
-
-export const AutoCompleteExampleGroups = () => {
-  const [value, setValue] = useState<string | null>(null);
-  return (
-    <StoryBookExample className={cnDocsDecorator('Section')}>
-      <AutoComplete
-        type="text"
-        placeholder="Введите значение"
-        value={value}
-        items={items}
-        groups={groups}
-        onChange={({ value }) => setValue(value)}
-      />
-    </StoryBookExample>
-  );
-};
