+++ src/components/AutoCompleteCanary/__stand__/AutoComplete.stand.mdx
@@ -1,0 +1,11 @@
+import { AutoCompleteExampleBasic } from './examples/AutoCompleteExampleBasic/AutoCompleteExampleBasic';
+
+```tsx
+import { AutoComplete } from '@consta/uikit/AutoCompleteCanary';
+```
+
+Поле ввода с подсказками. Позволяет как написать текст вручную так и использовать
+предустановленные подсказки. Этот компонент выглядит как [TextField](##LIBS.LIB.STAND/lib:uikit/stand:components-textfield-stable),
+но возможностей больше.
+
+<AutoCompleteExampleBasic />
