+++ src/components/AutoCompleteCanary/__stand__/AutoComplete.stand.tsx
@@ -1,0 +1,14 @@
+import { createStand } from '##/stand/standConfig';
+
+export default createStand({
+  title: 'AutoComplete',
+  id: 'AutoComplete',
+  group: 'components',
+  description: 'Поле для ввода текста с подсказками.',
+  version: '1.0.0',
+  status: 'canary',
+  // sandbox: 'textfield-0tb6vl',
+  figma:
+    'https://www.figma.com/embed?embed_host=share&url=https%3A%2F%2Fwww.figma.com%2Ffile%2Fv9Jkm2GrymD277dIGpRBSH%2FConsta-UI-Kit%3Fnode-id%3D96749%253A188498',
+  order: 10,
+});
