+++ src/components/AutoCompleteCanary/__stand__/AutoComplete.variants.css
@@ -1,0 +1,9 @@
+.AutoCompleteVariants {
+  &.canary--AutoComplete {
+    position: absolute;
+    top: var(--space-l);
+    width: 100%;
+    max-width: 400px;
+    padding: var(--space-m);
+  }
+}
