+++ src/components/AutoCompleteCanary/helpers.ts
